requirements.txt
Django==4.2.17
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.9
dj-database-url==2.2.0
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.3
django-cors-headers==4.3.0
